Ignore alt+tab in defaultTabEventHandler
As discussed on
https://groups.google.com/a/chromium.org/d/topic/blink-dev/9NRKOaJhElQ/discussion. 

The Chrome OS window manager helpfully (?) forwards all key press
events, including alt+tab, when an app is full screen with the
expectation that the web contents will ignore the event and send it back
to the window manager if it is not interested in it.  However, this
never happens in practice because the event gets default handled in
blink if it is ignored by the web contents.

The reason this isn't a problem for Chrome is that it checks if the key
press event is a special key combination and handles it directly without
passing it on to the web contents if it is (see
BrowserView::PreHandleKeyboardEvent).  Chrome also explicitly skips the
check if the current Browser object is an app, which is why this problem
only manifests itself for apps in full screen mode.

We can't change the check in BrowserView::PreHandleKeyboardEvent because
the current behavior was put in so that Chrome Remote Desktop users
would be able to forward alt+tab to their remote machines and adding the
check for apps would break CRD.  An alternative is have apps detect
alt+tab and call preventDefault() on it so that the event falls through
to the window manager.  This is also not great because we don't want
apps to deal with the details of the chrome os window manager and using
preventDefault() to prevent the default browser action so that the
default window manager action can be taken instead is a little weird to
think about.

The only option left is to fix blink's default tab handler to ignore
alt+tab events.  On Mac OSX, alt+tab (or option+tab) is a shortcut whose
behavior is governed by a system-wide setting.  We don't actually
respect this setting but that is separate issue.  On all other
platforms, we can safely ignore alt+tab and allow the event to fall back
to the window manager.

BUG=436670

Review URL: https://codereview.chromium.org/759813003

git-svn-id: svn://svn.chromium.org/blink/trunk@186266 bbb929c8-8fbe-4397-9dbb-9b2b20218538