IDL: Null values should be converted for non-nullable dictionary members (re-land)
This is the second attempt to land reverted CL (https://codereview.chromium.org/765673005/)
with GN fix.

Original description:

The current implementation of the binding layer treats null values
as the same as undefined, but this behavior doesn't match what the
spec specified. According to the spec[1], when null is passed to
a dictionary member, we should do conversion unless the member is
nullable. For example, given a dictionary member and the type
of the member is DOMString, passing null should result in "null"
(of type string). This CL changes the binding's behavior to follow
the spec.

[1] http://heycam.github.io/webidl/#es-dictionary
BUG=321462

Review URL: https://codereview.chromium.org/771813002

git-svn-id: svn://svn.chromium.org/blink/trunk@186276 bbb929c8-8fbe-4397-9dbb-9b2b20218538