Use invalidation instead of re-attach in HTMLDetailsElement.
When we open/close the details element we need to invaldiate the
RenderDetailsMarker in order to turn the arrow. Currently, we do this by
re-attaching the Summary section.

Instead, we can set the RenderDetailsMarker as needing invalidation and let
the invalidation after composting update system handle the work for us.

BUG=438642

Review URL: https://codereview.chromium.org/778703002

git-svn-id: svn://svn.chromium.org/blink/trunk@186445 bbb929c8-8fbe-4397-9dbb-9b2b20218538