Oilpan: Remove Heap::s_inGC
This CL fixes threading races on Heap::s_inGC by removing Heap::s_inGC. Given that Heap::s_inGC==true is almost identical to ThreadState::m_gcState==GCRunning, this CL makes the replacement.

The only downside of this CL is that we need to remove a couple of ASSERT(isInGC()) from methods that are executed in preGC(). This is because ThreadState::m_gcState is set to GCRunning in preGC() and thus we cannot assert that ThreadState::m_gcState is GCRunning or not in preGC(). (We can introduce another state like PreGCRunning to keep the assert, but I don't think it's worth doing.)

BUG=438565

Review URL: https://codereview.chromium.org/773073005

git-svn-id: svn://svn.chromium.org/blink/trunk@186479 bbb929c8-8fbe-4397-9dbb-9b2b20218538