Don't return false for RenderStyle::operator== while fonts load.
This caused an assert for baseRenderStyle in ActiveAnimations to hit. When
a font finishes loading, we trigger a full recalc of Document. That recalc
relies on getting a visual diff to trigger re-layout and paint and was
triggered by the Font::operator==. Instead, we remove the false return from
that operator and explicitly check if font loading status changed in
RenderStyle::diffNeedsFullLayoutAndPaintInvalidation().

R=eae@chromium.org
BUG=433765

Review URL: https://codereview.chromium.org/754743004

git-svn-id: svn://svn.chromium.org/blink/trunk@186572 bbb929c8-8fbe-4397-9dbb-9b2b20218538