Ensure floats are placed correctly when they exceed line width.
The bug report is for an ASSERT but there is a layout issue here too. When a float
exceeds the width of the line during line layout (and we've already placed a float
that exceeds the width of the line on it) we wait until the line break before placing
it. This is fine if there's just one - but if there's a bunch of them only the first
one will get positioned correctly. Refactor the float postioning code to place all
trailing floats in this situation.

BUG=366052

Review URL: https://codereview.chromium.org/764623003

git-svn-id: svn://svn.chromium.org/blink/trunk@186720 bbb929c8-8fbe-4397-9dbb-9b2b20218538