Service Worker: Use the error message from the embedder when throwing exceptions
WebServiceWorkerError has a message field that was previously ignored.
It's sometimes populated on browser-side.

I considered making the error message the "unsanitatedMessage" of the
DOMException, but I couldn't succeed in getting the message to show
up in console. It seems fine to just use it directly, as it doesn't
come from user input.

BUG=439810

Review URL: https://codereview.chromium.org/799503002

git-svn-id: svn://svn.chromium.org/blink/trunk@187219 bbb929c8-8fbe-4397-9dbb-9b2b20218538