Protect image's element while performing ops that could cause it to be gc'd
When an asyncronous image load completes it is possible that the element that
owns the load has been removed from the DOM. Now that image load can result in us
creating alt content with shadow DOM we can trigger mutation events, and these
may cause the element to be gc'd while we're still using it.

To ensure we hold on to the element while creating alt content, protect it with
a RefPtr until we are done.

BUG=440834

Review URL: https://codereview.chromium.org/800963002

git-svn-id: svn://svn.chromium.org/blink/trunk@187288 bbb929c8-8fbe-4397-9dbb-9b2b20218538