Make ImageFrame be transparent by default
Previously, ImageFrame was initialized to not have alpha, which means
that it would be considered to be opaque by other parts of the system.
Instead, initialize it to think it has alpha when it isn't known (which
the rest of the class also does).

This fix prevents flakiness in painting where loading caused raciness
with whether or not a transparent background image would cause a
background color to not get painted because of different values being
returned from backgroundHasOpaqueTopLayer.

BUG=442991

Review URL: https://codereview.chromium.org/814573003

git-svn-id: svn://svn.chromium.org/blink/trunk@187404 bbb929c8-8fbe-4397-9dbb-9b2b20218538